# demo/hello/hello-api-eda/target/generated-sources/src/main/java/com/sqli/pbousquet/helloapi/generated/api/server/HelloApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sqli.pbousquet.helloapi.generated.api.server;

import com.sqli.pbousquet.helloapi.generated.api.model.ApiErrorResponse;
import com.sqli.pbousquet.helloapi.generated.api.model.HelloDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Controller
@Tag(name = "Hello", description = "Hello API")
public interface HelloApi {

    default HelloApiDelegate getDelegate() {
        return new HelloApiDelegate() {};
    }

    /**
     * GET /api/v1/hello/{name} : Saluer une personne en particulier
     * Retourne un message de salutation personalisé.
     *
     * @param name Nom de la personne à saluer (required)
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "helloWithName",
        summary = "Saluer une personne en particulier",
        tags = { "Hello" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = HelloDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/hello/{name}",
        produces = { "application/json" }
    )
    default ResponseEntity<HelloDto> helloWithName(
        @Pattern(regexp = "^[a-zA-Z ,.'-]+$") @Size(min = 2, max = 25) @Parameter(name = "name", description = "Nom de la personne à saluer", required = true) @PathVariable("name") String name
    ) {
        return getDelegate().helloWithName(name);
    }


    /**
     * GET /api/v1/hello : Saluer le monde
     * Retourne un message de salutation générique.
     *
     * @return OK (status code 200)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "helloWorld",
        summary = "Saluer le monde",
        tags = { "Hello" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = HelloDto.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/hello",
        produces = { "application/json" }
    )
    default ResponseEntity<HelloDto> helloWorld(
        
    ) {
        return getDelegate().helloWorld();
    }

}